# openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /schedule:
    post:
      summary: Add Schedule
      operationId: add_schedule_schedule_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleModel'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleIdModel'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Get Schedule For User By Id
      operationId: get_schedule_for_user_by_id_schedule_get
      parameters:
      - name: schedule_id
        in: query
        required: true
        schema:
          type: integer
          exclusiveMinimum: 0
          title: Schedule Id
      - name: user_id
        in: query
        required: true
        schema:
          type: integer
          exclusiveMinimum: 0
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/FullScheduleResponse'
                - $ref: '#/components/schemas/MessageResponse'
                title: Response Get Schedule For User By Id Schedule Get
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /schedules:
    get:
      summary: Get Schedules
      operationId: get_schedules_schedules_get
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: integer
          exclusiveMinimum: 0
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulesListResponse'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /next_takings:
    get:
      summary: Get Schedule For Period
      operationId: get_schedule_for_period_next_takings_get
      parameters:
      - name: user_id
        in: query
        required: true
        schema:
          type: integer
          exclusiveMinimum: 0
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DynamicSchedulesResponse'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    DynamicSchedulesResponse:
      additionalProperties:
        items:
          type: string
        type: array
      type: object
      title: DynamicSchedulesResponse
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorResponse
    FullScheduleResponse:
      properties:
        schedule_id:
          type: integer
          title: Schedule Id
        medicine:
          type: string
          title: Medicine
        time_list:
          items:
            type: string
          type: array
          title: Time List
      type: object
      required:
      - schedule_id
      - medicine
      - time_list
      title: FullScheduleResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
    ScheduleIdModel:
      properties:
        schedule_id:
          type: integer
          title: Schedule Id
      type: object
      required:
      - schedule_id
      title: ScheduleIdModel
    ScheduleModel:
      properties:
        user_id:
          type: integer
          title: User Id
        medicine:
          type: string
          title: Medicine
        frequency:
          type: integer
          title: Frequency
        duration_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Days
      type: object
      required:
      - user_id
      - medicine
      - frequency
      title: ScheduleModel
    SchedulesListResponse:
      properties:
        active_schedule_ids:
          items:
            type: integer
          type: array
          title: Active Schedule Ids
      type: object
      required:
      - active_schedule_ids
      title: SchedulesListResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError